% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_management_fns.R
\name{add_cpiametadata}
\alias{add_cpiametadata}
\title{Add CPIA Variable Metadata to a Dataset}
\usage{
add_cpiametadata(
  df,
  indicator,
  variable,
  source,
  var_name,
  var_description,
  var_description_short,
  theoretical_range
)
}
\arguments{
\item{df}{A `data.frame` to which metadata attributes will be added.}

\item{indicator}{A `character` string identifying the indicator code
(e.g., `"bs_bti_q15_1"` or `"vdem_core_v2clacjstm"`).}

\item{variable}{the cpia criteria specific question (e.g, 12a-c, 15b, 16a etc)}

\item{source}{A `character` string indicating the data source
(e.g., `"CLIAR"`, `"V-Dem"`, `"IBP"`).}

\item{var_name}{A `character` string providing the variable’s short name
(e.g., `"Efficient use of assets"`, `"Open budget index"`).}

\item{var_description}{A `character` string giving a detailed explanation
of what the variable measures.}

\item{var_description_short}{A concise summary of the variable’s meaning
suitable for dashboards or reports.}

\item{theoretical_range}{the theoretical range for the dataset e.g (c(0, 100))}
}
\value{
The input `data.frame` with the following additional attributes:
\describe{
  \item{`indicator`}{Indicator code used to identify the data source.}
  \item{`variable`}{CPIA variable code corresponding to a question or subquestion.}
  \item{`source`}{Source organization or dataset from which the variable is derived.}
  \item{`var_name`}{Short variable label.}
  \item{`var_description`}{Full variable definition or interpretation.}
  \item{`var_description_short`}{Concise summary of the variable meaning.}
  \item{`theoretical_range`}{The range of the raw}
}
}
\description{
Attaches standardized CPIA metadata attributes to a dataset.
This function is designed to tag each CPIA-related dataset with descriptive information
such as the question, subquestion, variable source, and indicator definitions.
It ensures that all datasets contributing to the CPIA composite data system
carry consistent metadata attributes for traceability and documentation.
}
\examples{
df <- data.frame(value = rnorm(5))
df_meta <- add_cpiametadata(
  df,
  indicator = "bs_bti_q15_1",
  variable = "q15a",
  source = "CLIAR",
  var_name = "Efficient use of assets",
  var_description = "To what extent does the government make efficient use of assets?",
  var_description_short = "Measures the extent to which the government makes efficient use of assets."
  theoretical_range = c(0, 1)
)
attributes(df_meta)

}
